Generation Completed
Final changes on generator includes updating how the DR are received (input DR only).
Now, any generated stage creates a DR if it's an output, according to its naming
conventions and receives the input DR name, in order to access it.

Also DiffMaskComp updated to be compatible to the current stage generator
strategy. Still need to think on how the workflow final output should be
handled.

Currently the code is compiling. No running tests were made so far.