Inputs, outputs and stages retreived correctly.
All three functionalities tested and working (so far...). The Argument and
PipelineComponentBase got a name and id field, needed for the generation of
the workflow. Those fields do not impact any other place of the old code
after the workflow generation.

Another possibility is to extend both those classes in order for the new
classes to have the missing needed fields. This idea was initially discarted
on the account of the length of the code needed for the new functionalities:
too little code for too many classes (including all subclasses of those classes).