Input parameter expansion
Implemented and tested an algorithm to expand the parameters combinations. This
means that for a list of parameters, each with one or more possible values, a new
list of parameters set is generated. A parameter set is a set of input parameters with
only one value, being the set enough to start the workflow execution. The list contains
all possible combinations of parameters' values.