Bug fixes
1. Corrected generated stages regarding RT/DR variables initialization,
2. Updated ArgumentBase to hold an io_type, representing if a variable is an
input or output. This will be used later to perform memory cleaning on arguments.
Serialization/deserialization methods updated to hold this state between messages.
3. Commented memory cleaning line PipelineComponentBase.cpp:42. This operation
resulted on premature deallocaion of reused arguments and it would possibly result
in a racing condition, on which if a reusable argument is read by a second task before
a first task deallocated it.