Bug fix:
It was possible for a reusable task to break dependencies when being reused.
E.g:    A__B__C__G
         \_D__E__H
            \_F__I
Given the above task dependency tree if a stage S' was to be merged to the
stage S (above), both of three tasks per workflow, and tasks A, D and C were
said to be reusable by S' tasks the correct merged stage should be:

        A____B__C__G
         \___S'2__S'3__S'4
          \__D__E__H
              \_F__I

And instead the algorithm was retunning:

        A__B___C___G
         \      \__S'4
          \_D__E___H
             \_F___I

Breaking the dependency of C to B with D.

The algorithm was fixed and tested successfully with 4 parameters with 2
values each, on -np 7.