Stoer-Wagner min-cut algorithm implemented.
As far as it was tested it doesn't present any bug.
In order to compile it the boost lib is required.
The output isn't still finished (i.e outputs on stdout
only). Also, the cut weight ins't implemented yet, so
the cut value on the output is a miscalculated value.

The min-cut algorithm isn't integrated on the system yet.