Recursive montecarlo implementation created.
This implementation goal is to outperform the full recursive cutting
algorithm execution time. The solution, however, won't be the 'best'
solution (as it was especulated that the full recursive algorithm had
such performance).

In order to test with larger test sets (71 stages on 5 workers) the
memory leak problem (intertask cvMat's) was re-fixed, thereby,
breaking the execution stage.