Fixed the Diff memory leak
Leak was due to the non deletion of the DDR2Ds of DiffMaskComp.cpp:run() whenever
one of the two DDR2D was NULL. This was due to some bad stage merging, which
caused the input arguments to have the wrong (reused stage) dependency ID, thus
making such arguments unavailable.

The arguments' IDs were fixed.